+++ src/color/color.cpp
@@ -1,7 +1,7 @@
 #include <color.h>
 
-Color::Color(double r, double g, double b) : r(r), g(g), b(b) {}
+Color::Color(float r, float g, float b) : r(r), g(g), b(b) {}
 
-double Color::getR() const { return r; }
-double Color::getG() const { return g; }
-double Color::getB() const { return b; }
+float Color::getR() const { return r; }
+float Color::getG() const { return g; }
+float Color::getB() const { return b; }
